ajout de l'activity report

+++ api/src/main/java/fr/orleans/sig/repository/CorbeillesRepository.java
@@ -3,5 +3,5 @@ package fr.orleans.sig.repository;
 import fr.orleans.sig.model.geo.Corbeilles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CorbeillesRepository extends JpaRepository<Corbeilles, Long> {
+public interface CorbeillesRepository extends JpaRepository<Corbeilles, Integer> {
 }
