+++ api/src/main/java/fr/orleans/sig/repository/DechetsRepository.java
@@ -3,5 +3,5 @@ package fr.orleans.sig.repository;
 import fr.orleans.sig.model.geo.DechetsPav;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DechetsRepository extends JpaRepository<DechetsPav, Long> {
+public interface DechetsRepository extends JpaRepository<DechetsPav, Integer> {
 }
