+++ api/src/main/java/fr/orleans/sig/repository/SanitairesRepostiory.java
@@ -4,5 +4,5 @@ import fr.orleans.sig.model.geo.Sanitaires;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface SanitairesRepostiory extends JpaRepository<Sanitaires, Long> {
+public interface SanitairesRepostiory extends JpaRepository<Sanitaires, Integer> {
 }
